+++ app/Http/Controllers/Admin/IntroController.php
@@ -8,6 +8,7 @@ use App\Http\Services\intro\introService;
 use App\Models\c;
 use App\Models\intro;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
 class IntroController extends Controller
@@ -25,6 +26,8 @@ class IntroController extends Controller
      */
     public function index()
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         return view('admin.intros.list', [
             'title' => 'Danh sách giới thiệu',
             'data' => $this->introService->get()
@@ -38,6 +41,8 @@ class IntroController extends Controller
      */
     public function create()
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         return view('admin.intros.add', [
             'title' => 'tạo trang giới thiệu'
         ]);
@@ -51,6 +56,8 @@ class IntroController extends Controller
      */
     public function store(createIntroRequest $request)
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         $row = $this->introService->countRow();
         if (count($row) != 0) {
             Session::flash('error', 'Trang giới đã tồn tại bạn không thể tạo thêm nữa');
@@ -73,6 +80,8 @@ class IntroController extends Controller
      */
     public function show(c $c)
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         //
     }
 
@@ -84,6 +93,8 @@ class IntroController extends Controller
      */
     public function edit(intro $id)
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         return view('admin.intros.edit', [
             'title' => 'tạo trang giới thiệu',
             'data' => $id
@@ -99,6 +110,8 @@ class IntroController extends Controller
      */
     public function update(createIntroRequest $request, intro $id)
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         $resuit = $this->introService->update($id, $request);
         if ($resuit) {
             return redirect('/admin/intro/list');  
@@ -115,6 +128,8 @@ class IntroController extends Controller
      */
     public function destroy(Request $request)
     {
+        if (Auth::user()->level != 1) return redirect()->route('admin');
+        
         $resuit = $this->introService->destroy($request);
         if ($resuit) {
             return response()->json(['error' => false, 'message' => 'Đã xóa giới thiệu thành công']);
